Export additional utility functions
