+++ src/support/devs.ts
@@ -1,0 +1,4 @@
+export type Progress = {
+  x: number;
+  y: number;
+};
